requirements.txt
streamlit
opencv-python
scikit-image
scikit-learn
joblib
numpy
pillow
